+++ src/digsim/app/model/_model_objects.py
@@ -4,8 +4,6 @@
 """Handle objects in the model"""
 
 from digsim.circuit import Circuit
-from digsim.circuit.components.atoms import DigsimException
-from digsim.storage_model import AppFileDataClass, ModelDataClass
 
 from ._model_components import ModelComponents
 from ._model_new_wire import NewWire
@@ -69,37 +67,9 @@ class ModelObjects:
             self._app_model.model_changed()
         self._app_model.sig_delete_wires.emit()
 
-    def model_to_circuit(self, model_dc, circuit_folder):
-        if isinstance(model_dc, AppFileDataClass):
-            # Loaded model
-            dc = ModelDataClass.from_app_file_dc(model_dc)
-        else:
-            dc = model_dc
-
-        try:
-            # Create circuit
-            exception_str_list = self.circuit.from_dataclass(
-                dc.circuit,
-                circuit_folder,
-                component_exceptions=False,
-                connect_exceptions=False,
-            )
-            # Add component positions
-            self.components.add_gui_positions(dc.gui)
-        except DigsimException as exc:
-            self._app_model.sig_error.emit(f"Circuit error: {str(exc)}")
-            return exception_str_list
-        return exception_str_list
-
-    def circuit_to_model(self, circuit_folder):
-        model_dc = ModelDataClass(
-            circuit=self.circuit.to_dataclass(circuit_folder), gui=self.components.get_gui_dict()
-        )
-        return model_dc
-
     def _restore_state(self, model_dc):
         self.clear()
-        exception_str_list = self.model_to_circuit(model_dc, None)
+        exception_str_list = self._app_model.model_to_circuit(model_dc)
         self._app_model.model_init()
         self._app_model.model_changed()
         if len(exception_str_list) > 0:
@@ -115,7 +85,7 @@ class ModelObjects:
 
     def push_undo_state(self, clear_redo_stack=True):
         """Push undo state to stack"""
-        self._undo_stack.append(self.circuit_to_model("/"))
+        self._undo_stack.append(self._app_model.circuit_to_model())
         if clear_redo_stack:
             self._redo_stack = []
             self._app_model.sig_control_notify.emit()
@@ -128,7 +98,7 @@ class ModelObjects:
 
     def push_redo_state(self):
         """Push redo state to stack"""
-        self._redo_stack.append(self.circuit_to_model("/"))
+        self._redo_stack.append(self._app_model.circuit_to_model())
 
     def undo(self):
         """Undo to last saved state"""