+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## vx.x.x
+
+## v1.0.0
  - Bugfix: Fix bad signal path
  - Bugfix: Disconnect signal before GUI component removal
  - Cleanup path handling